Don't flush terminal input when querying checking window size
As reported in https://github.com/MidnightCommander/mc/issues/4789,
after Midnight Commander starts a shell, it writes some characters
to the shells stdin, to initialize it.

Since mc doesn't know when shell initialization
(e.g. .bashrc), it doesn't wait for that.
Hence mc's commands race with bashrc.

This surfaces reliably when running fastfetch from .bashrc with a *.png
logo.  Writing the logo involves a call to ffDetectTerminalSize(),
which queries /dev/tty.

It first flushes (TCSAFLUSH) the input since a0705197 (IO: disallow
input echo, 2024-07-31), which removes the input queued by mc.

Then it reads from the TTY 1024 bytes at a time.
This is also bad because it might swallow key presses
made by the user.

So maybe
1. Don't flush.
2. Reduce the likelihood of swallowing users's key presses by reading
   only one byte at a time.

In theory this might result in us reading wrong values if the user
types very quickly at startup (possibly synthesized, as is the case
with mc).